[ConfigUI] Use file to save backend configuration

+++ configuration-ui/src/api/index.php
@@ -4,17 +4,25 @@ require 'vendor/autoload.php';
 require 'bidders.php';
 require 'publishers.php';
 
-$app = new \Slim\Slim();
-$app->config('debug', false);
+if (!file_exists('config/config.php'))
+	die('config/config.php is not found');
+$config = array();
+include_once("config/config.php");
 
-$bidders = new Bidders();
+$app = new \Slim\Slim();
+$app->config('debug', $config['debug']);
+
+$db = new RedMap\Drivers\MySQLiDriver('UTF8');
+$db->connect($config['db_user'], $config['db_password'], $config['db_name']);
+
+$bidders = new Bidders($db);
 $app->get('/bidders/', function () use ($app, $bidders) { display_result($app, $bidders->getAll($app)); });
 $app->get('/bidders/:id', function ($id) use ($app, $bidders) { display_result($app, $bidders->get($app, $id)); });
 $app->delete('/bidders/:id', function ($id) use ($app, $bidders) { $bidders->delete($app, $id); });
 $app->put('/bidders/:id', function ($id) use ($app, $bidders) { $bidders->put($app, $id); });
 $app->post('/bidders/', function () use ($app, $bidders) { $bidders->post($app); });
 
-$publishers = new Publishers();
+$publishers = new Publishers($db);
 $app->get('/publishers/', function () use ($app, $publishers) { display_result($app, $publishers->getAll($app)); });
 $app->get('/publishers/:id', function ($id) use ($app, $publishers) { display_result($app, $publishers->get($app, $id)); });
 $app->delete('/publishers/:id', function ($id) use ($app, $publishers) { $publishers->delete($app, $id); });
